refactor(author): перенести сервис по авторам в authors-service
МЕТА ДАННЫЕ: #3
